Fix small error in the user guide for OptimTraj
There were two instances in the user guide where the order of the
equality and inequality constraints were swapped. This commit
fixes the two mistakes.

This commit also removes some whitespace in the .tex files and
adds a reference to my primary tutorial paper at the end of the
user guide.